+++ src/http-server/index.js
@@ -2,7 +2,7 @@
 
 import {digg} from "diggerize"
 import EventEmitter from "../utils/event-emitter.js"
-import {Logger} from "../logger.js"
+import Logger from "../logger.js"
 import Net from "net"
 import ServerClient from "./server-client.js"
 import WorkerHandler from "./worker-handler/index.js"