+++ src/routes/resolver.js
@@ -4,7 +4,7 @@ import {dirname} from "path"
 import {fileURLToPath} from "url"
 import fs from "fs/promises"
 import * as inflection from "inflection"
-import {Logger} from "../logger.js"
+import Logger from "../logger.js"
 import UploadedFile from "../http-server/client/uploaded-file/uploaded-file.js"
 import ensureError from "../utils/ensure-error.js"